+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased] - ReleaseDate
 
+### Fixed
+
+- Fix pager view action (regression in 3.1.2)
+
 ## [3.1.2] - 2025-05-30
 
 ### Changed